+++ src/pages/EditOperatorData.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import CreateLayout from "../components/Layout/CreateLayout";
-import EditOperatorDataForm from "../components/Form/EditOperatorDataForm/EditOperatorDataForm";
+import { EditOperatorDataForm } from "../components/Form/EditOperatorDataForm/EditOperatorDataForm";
 
 const EditOperatorData = () => (
   <CreateLayout breadcrumbLabel="Modifica dati" title="Dati operatore">
