+++ src/pages/EditProfile.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import CreateLayout from "../components/Layout/CreateLayout";
-import EditProfileForm from "../components/Form/EditProfileForm/EditProfileForm";
+import { EditProfileForm } from "../components/Form/EditOperatorDataForm/EditOperatorDataForm";
 
 const EditProfile = () => (
   <CreateLayout breadcrumbLabel="Modifica dati" title="Profilo">
